added function for overwriting assertions
